Restructure test files
Discussed in
https://github.com/knatten/cppquiz/pull/296#issuecomment-1487315818.
