+++ quiz/tests/system/test_directly_to_question.py
@@ -1,5 +1,5 @@
-from quiz.system_tests.system_test_case import SystemTestCase
-from quiz.test_helpers import create_questions
+from quiz.tests.system.system_test_case import SystemTestCase
+from quiz.tests.test_helpers import create_questions
 
 
 class DirectlyToQuestionTest(SystemTestCase):
