+++ quiz/tests/unit/test_auto_publish.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 from django.utils import timezone
 
 from quiz.models import Question
-from quiz.test_helpers import create_questions
+from quiz.tests.test_helpers import create_questions
 
 
 class AutoPublishTest(TestCase):
