+++ quiz/tests/unit/test_remove_section_numbers.py
