+++ quiz/tests/unit/test_session.py
@@ -3,7 +3,7 @@ import unittest
 import mock
 
 from quiz.game_data import UserData, save_user_data
-from quiz.test_helpers import create_questions
+from quiz.tests.test_helpers import create_questions
 
 
 def create_session_with_answers_to(questions):
